requirements.txt
chromadb==0.4.22
openai==1.12.0
python-dotenv==1.0.1